Move copy test to seperate file
